Create README.md
feat: Add profile README for GitHub profile

This commit adds a new profile README file to the top of the GitHub profile page. The README includes a professional summary, achievements, vision, hobbies, and contact information.

- Professional Summary: Highlights of my 6+ years of experience in Software Testing and Quality Assurance.
- Achievements: Client appreciation and Pillar of the Month recognition.
- Vision: Commitment to enhancing IT technology and Quality Compliance.
- Hobbies: Interests in reading, watching movies, and creating DIY artistic decors.
- Contact Information: Email and LinkedIn profile for professional connections.

This README aims to provide visitors with an overview of my professional background and personal interests.